Remove full page cover
Previously when the model is running the full page is
covered and user is not able to operate the page, which
is a not good user experience.

This PR fixes the issue by disable buttons click during
model running to avoid the unexpected errors when switching
the model before last model running is done.