Refactor the Facial Landmark Detection example to fetch in parallel
This model separates each constant value into a separate NumPy file
which needs to be fetched. This change refactors the graph construction
so that these fetches can happen in parallel rather than waiting for
each fetch to complete before building that graph layer.

This is an experiment to determine whether it would be more ergonomic
for the graph builder function to accept promises so that this kind of
code can be written without the careful placement of awaits demonstrated
here.