Add NNotepad sample
It is not yet linked from the main README or index page as there is
significant integration/cleanup work to do. But it is useful as-is
for WebNN implementers.

A bespoke test harness is included. Documentation visible in the
sample is generated from the README.md (run `make` to update).