[NNotepad] Fix serialization of non-tensor arrays
Most MLGraphBuilder method arguments are operands, but sometimes
they're arrays or numbers. There was a mechanism in place to determine
(via a table) whether a given method's argument was an operator or
not, but that wasn't enough to distinguish an array of numbers (like
reshape() needs) from an array of operands (which concat() needs).

Rework the lookup from returning a boolean to returning an argument
type, and adjust serialization appropriately.

Fixes #263