NNotepad: Pass buffers rather than views when building constants (#303)
When building constants, either a raw ArrayBuffer can be passed or a
view type like Float32Array. In the latter case the data type and view
type must match; e.g. if dataType is "int8", a Uint8Array must be
passed.

Chromium is in the process of rolling out native Float16Array support,
but currently the WebNN prototype implementation[1] requires passing a
Uint16Array if the data type is "float16". NNotepad utilizes this
workaround if native Float16Array is not present. But if it *is*
present, it passes a native Float16Array and Chromium throws.

Until things settle, have NNotepad's constant() builder just pass the
underlying buffer rather than view, which bypasses this check. This
change can be reverted at some point in the future, but is harmless to
continue using this workaround.

1: https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/renderer/modules/ml/webnn/ml_graph_utils.cc;l=92