NNotepad: Parse the IDL to determine argument types
To know how to parse argument and dictionary members - are they
tensors or simple types? - the parser asks a WebNNUtil function. This
function did the job by hardcoding a list of the special cases. This
requires careful maintenance.

Update the code to bundle a copy of the WebIDL and parse it, instead.
This uses webidl2.js, which is the same library used by Web Platform
Tests for validating the IDL.

This will require updating the snapshot IDL as the spec evolves, but
that's less specialized than tweaking the special JSON data and can
potentially be automated.

Future work includes:

- Fetch the IDL from the Web Platform Tests repository, or some other
  CORS-friendly canonical resource.

- Use the parsed IDL to drive the Monaco editor, for better
  autocomplete.