NNotepad: Add ability to show additional tensors in the output (#311)
Adds `output()` helper which can be used to list an identifier (or
multiple identifiers) that will be shown in the results pane before
the result of the final expression or assignment.

A few notes:

- Ending with an `output()` is supported, as a convenience.

- The `output()` helper takes an identifier, not an expression. In the
  future this could change, either to taking an expression or showing
  the identifier name in the results pane.

- Mixing `split()` (which returns a list of tensors) and `output()` is
  probably a bit confusing, since split already gives you multiple
  tensors, you're gonna have to pay attention to how many are coming
  from each. Sorry!

Also fixes a minor issue with the call to asyncInit in js/tests.js -
it wasn't awaited, so if the first test case was a test with a call it
would fail.